+++ tango/executors/__init__.py
@@ -1,0 +1,4 @@
+"""
+Built-in :class:`~tango.executor.Executor` implementations.
+"""
+from .multicore_executor import MulticoreExecutor