+++ test_fixtures/package/steps.py
@@ -51,6 +51,54 @@ class AddNumbersStep(Step):
         return a_number + b_number
 
 
+@Step.register("sleep-print-maybe-fail")
+class SleepPrintMaybeFail(Step):
+    DETERMINISTIC = True
+    CACHEABLE = True
+
+    def run(self, string: str, seconds: int = 5, fail: bool = False) -> str:  # type: ignore
+        time.sleep(seconds)
+        self.logger.info(f"Step {self.name} is awake.")
+        print(string)
+        if fail:
+            raise RuntimeError("Step had to fail!")
+        return string
+
+
+@Step.register("logging-step")
+class LoggingStep(Step):
+    DETERMINISTIC = True
+    CACHEABLE = True
+
+    def run(self, string: str, num_log_lines: int = 50) -> str:  # type: ignore
+        for i in Tqdm.tqdm(list(range(num_log_lines)), desc="log progress"):
+            time.sleep(0.1)
+            self.logger.info(f"{i} - {string}")
+        return string
+
+
+@Step.register("make_number")
+class MakeNumber(Step):
+    DETERMINISTIC = True
+    CACHEABLE = True
+
+    def run(self, what_number: int) -> int:  # type: ignore
+        return what_number
+
+
+@Step.register("store_number_in_file")
+class StoreNumberInFile(Step):
+    DETERMINISTIC = True
+    CACHEABLE = False
+
+    def run(self, number: int, file_name: str) -> None:  # type: ignore
+        # Note: this is only for testing if the uncacheable step
+        # ran in the multicore setting. Normally, a step like this
+        # would be marked as CACHEABLE.
+        with open(file_name, "w") as file_ref:
+            file_ref.write(str(number))
+
+
 @Step.register("multiprocessing_step")
 class MultiprocessingStep(Step):
     """